Added explicit casts for float to int in the ToneDetector since that casued warnings on some compilers.
git-svn-id: svn://svn.code.sf.net/p/svxlink/svn/trunk@2039 f8843c6b-282e-0410-804b-c0d5895947d2

+++ src/svxlink/trx/ToneDetector.cpp
@@ -186,7 +186,8 @@ void ToneDetector::setDetectionDelay(int delay_ms)
 {
   if (delay_ms > 0)
   {
-    det_delay = ceil(delay_ms * INTERNAL_SAMPLE_RATE / (block_len * 1000.0));
+    det_delay = static_cast<int>(
+		ceil(delay_ms * INTERNAL_SAMPLE_RATE / (block_len * 1000.0)));
   }
   else
   {
@@ -205,7 +206,8 @@ void ToneDetector::setGapDelay(int delay_ms)
 {
   if (delay_ms > 0)
   {
-    gap_delay = ceil(delay_ms * INTERNAL_SAMPLE_RATE / (block_len * 1000.0));
+    gap_delay = static_cast<int>(
+		ceil(delay_ms * INTERNAL_SAMPLE_RATE / (block_len * 1000.0)));
   }
   else
   {
